requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
pandas
opentelemetry-exporter-gcp-trace
